All repositories are now functional

+++ spd-domain/src/main/java/br/com/spd/domain/AvaliacaoMat2.java
@@ -13,7 +13,7 @@ import javax.persistence.*;
 @NamedQuery(name="AvaliacaoMat2.findAll", query="SELECT a FROM AvaliacaoMat2 a")
 public class AvaliacaoMat2 implements Serializable {
 	private static final long serialVersionUID = 1L;
-	private Integer codavaliacao;
+	private Long codavaliacao;
 	private String artCorpo;
 	private String artCria;
 	private String artMateriais;
@@ -52,11 +52,11 @@ public class AvaliacaoMat2 implements Serializable {
 
 
 	@Id
-	public Integer getCodavaliacao() {
+	public Long getCodavaliacao() {
 		return this.codavaliacao;
 	}
 
-	public void setCodavaliacao(Integer codavaliacao) {
+	public void setCodavaliacao(Long codavaliacao) {
 		this.codavaliacao = codavaliacao;
 	}
